feat: systemd socket
Created a systemd socket unit will be responsible to listen the
connection, and when something connects to it, the daemon will
initialize the service.

Hooks were also updated to have a smooth transition between versions
installations to avoid any downtime.

Additionally, Golang code was updated to read the socket created, and
deal with service shutdown.